refactor the request method
