+++ src/libs/trpc/client/edge.ts
@@ -1,12 +1,18 @@
 import { createTRPCClient, httpBatchLink } from '@trpc/client';
 import superjson from 'superjson';
 
+import { isDesktop } from '@/const/version';
 import type { EdgeRouter } from '@/server/routers/edge';
 import { withBasePath } from '@/utils/basePath';
+
+import { fetchWithDesktopRemoteRPC } from './helpers/desktopRemoteRPCFetch';
 
 export const edgeClient = createTRPCClient<EdgeRouter>({
   links: [
     httpBatchLink({
+      fetch: isDesktop
+        ? (input, init) => fetchWithDesktopRemoteRPC(input as string, init)
+        : undefined,
       headers: async () => {
         // dynamic import to avoid circular dependency
         const { createHeaderWithAuth } = await import('@/services/_auth');
