+++ src/utils/fetch/index.ts
@@ -1,2 +1,4 @@
 export * from './fetchSSE';
+export * from './headers';
 export * from './parseError';
+export * from './request';
